+++ engine/classes/Elgg/I18n/InvalidLocaleException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Elgg\I18n;
+
+/**
+ * Indicated an invalid Locale was given
+ *
+ * @access private
+ */
+class InvalidLocaleException extends \InvalidArgumentException {}