+++ mod/custom_index/languages/pt_BR.php
@@ -1,8 +1,0 @@
-<?php
-return array(	
-	'custom:bookmarks' => "Últimos favoritos",
-	'custom:groups' => "Últimas comunidades",
-	'custom:files' => "Últimos arquivos",
-	'custom:blogs' => "Últimas mensagens dos blogs",
-	'custom:members' => "Membros mais recentes",
-);