+++ mod/groups/languages/ja.php
@@ -53,11 +53,11 @@ return array(
 	'groups:widget:group_activity:content:noselect' => 'グループを選択するにはこのウィジェットを編集してください',
 
 	'groups:noaccess' => 'グループへのアクセスを許可しない',
-	'groups:permissions:error' => 'あなたの権限ではこれはできません',
 	'groups:ingroup' => 'グループ内で',
 	'groups:cantcreate' => 'あなたは、グループを作成することができません。管理者のみ作成できます。',
 	'groups:cantedit' => 'このグループを編集できません。',
 	'groups:saved' => 'グループを保存しました',
+	'groups:save_error' => 'Group could not be saved',
 	'groups:featured' => 'クローズアップ',
 	'groups:makeunfeatured' => 'クローズアップをやめる',
 	'groups:makefeatured' => 'クローズアップに登録する',
@@ -81,30 +81,6 @@ return array(
 	'groups:search_in_group' => "このグループ内を検索",
 	'groups:acl' => "グループ: %s",
 
-	'discussion:topic:notify:summary' => '新しい議題「%s」が投稿されました',
-	'discussion:topic:notify:subject' => '新着議題: %s',
-	'discussion:topic:notify:body' =>
-'%s さんは、グループ %s に新しい議題を投稿しました:
-
-題名: %s
-
-%s
-
-この議題を閲覧したり議題に返答するには:
-%s
-',
-
-	'discussion:reply:notify:summary' => '議題 %s に対して新しい返答がありました。',
-	'discussion:reply:notify:subject' => '新着返答:  議題 %s',
-	'discussion:reply:notify:body' =>
-'%1$s さんがグループ %3$s の議題「%2$s」に返答しました:
-
-%4$s
-
-この会議を閲覧したり議題に返答するには:
-%5$s
-',
-
 	'groups:activity' => "グループのうごき",
 	'groups:enableactivity' => 'グループのうごきを有効にする',
 	'groups:activity:none' => "グループのうごきはありません。",
@@ -115,11 +91,6 @@ return array(
 	'groups:requests:none' => '現在、会員リクエストはありません。',
 
 	'groups:invitations:none' => '現在、グループへの招待はありません。',
-
-	'item:object:groupforumtopic' => "議題",
-	'item:object:discussion_reply' => "議論の返答",
-
-	'groupforumtopic:new' => "新規投稿",
 
 	'groups:count' => "グループ数",
 	'groups:open' => "オープングループ",
@@ -146,69 +117,10 @@ return array(
 	/**
 	 * Group tools
 	 */
-	'groups:enableforum' => 'グループ会議の利用',
 	'groups:lastupdated' => '最終更新 %s(%s さん)',
 	'groups:lastcomment' => '最新コメント %s(%s さん)',
 
-	/**
-	 * Group discussion
-	 */
-	'discussion' => '会議室',
-	'discussion:add' => '議題を追加',
-	'discussion:latest' => '最新の話題',
-	'discussion:group' => 'グループ会議',
-	'discussion:none' => '議論はありません',
-	'discussion:reply:title' => '%s さんからの返答',
-
-	'discussion:topic:created' => '議題を作成しました',
-	'discussion:topic:updated' => '議題を更新しました',
-	'discussion:topic:deleted' => '議題を削除しました',
-
-	'discussion:topic:notfound' => '議題は見つかりませんでした',
-	'discussion:error:notsaved' => 'この議題を保存できませんでした',
-	'discussion:error:missing' => 'タイトルとメッセージは必須項目です',
-	'discussion:error:permissions' => 'あなたには、このアクションを行う権限がありません',
-	'discussion:error:notdeleted' => '議題を削除できませんでした。',
-
-	'discussion:reply:edit' => '返答を編集',
-	'discussion:reply:deleted' => '返答を削除しました。',
-	'discussion:reply:error:notfound' => '議論の返答は見つかりませんでした。',
-	'discussion:reply:error:notfound_fallback' => "申し訳ありません。お探しの返答は見つかりませんでした。オリジナルの会議の議論にご案内します。",
-	'discussion:reply:error:notdeleted' => '返答を削除することができませんでした。',
-
-	'discussion:search:title' => '議題「 %s 」に返答する',
-	
 	'admin:groups' => 'グループ',
-
-	'reply:this' => '返答する',
-
-	'group:replies' => '返答',
-	'groups:forum:created' => '%s（返答 %d 件）を作成しました',
-	'groups:forum:created:single' => '%s（返答 %d 件）を作成しました',
-	'groups:forum' => 'グループ会議',
-	'groups:addtopic' => '新規議題作成',
-	'groups:forumlatest' => '最新のやりとり',
-	'groups:latestdiscussion' => '最新のやりとり',
-	'groupspost:success' => 'コメントを投稿しました。',
-	'groupspost:failure' => '返答を保存スル際に問題が発生しました。',
-	'groups:alldiscussion' => '最新のやりとり',
-	'groups:edittopic' => '議題の編集',
-	'groups:topicmessage' => '議題メッセージ',
-	'groups:topicstatus' => '議題の状態',
-	'groups:reply' => 'コメントの投稿',
-	'groups:topic' => '議題',
-	'groups:posts' => '投稿',
-	'groups:lastperson' => '最後に投稿した人',
-	'groups:when' => 'いつ',
-	'grouptopic:notcreated' => '議題はありません。',
-	'groups:topicclosed' => '議論終了',
-	'grouptopic:created' => '議題を作成しました。',
-	'groups:topicsticky' => 'スティッキー',
-	'groups:topicisclosed' => 'この議題は終了しました。',
-	'groups:topiccloseddesc' => 'この議題は終了しました。新しいコメントは受け付けていません。',
-	'grouptopic:error' => 'グループ議題が作成できません。システム管理者に問い合わせください。',
-	'groups:forumpost:edited' => "投稿を編集しました。",
-	'groups:forumpost:error' => "投稿を編集する際に問題が発生しました。",
 
 	'groups:privategroup' => 'このグループはクローズドグループです。参加希望を申請してください。',
 	'groups:notitle' => 'グループ作成にはグループ名が必要です。',
@@ -229,7 +141,6 @@ return array(
 	'groups:usernotinvited' => 'ユーザを招待できませんでした。',
 	'groups:useralreadyinvited' => 'ユーザはすでに招待済みです。',
 	'groups:invite:subject' => "%sさん、%s に招待されています。",
-	'groups:updated' => "最新の返答:(%s さんより)「 %s 」",
 	'groups:started' => "開始日: %s",
 	'groups:joinrequest:remove:check' => 'この招待リクエストを削除してよいですか？',
 	'groups:invite:remove:check' => 'この招待を破棄してもよろしいですか？',
@@ -263,9 +174,6 @@ return array(
 
 	'river:create:group:default' => '%s さんは、グループ「 %s 」を作成しました。',
 	'river:join:group:default' => '%s さんは、グループ「 %s 」に参加しました。',
-	'river:create:object:groupforumtopic' => '%s さんは、新しく議題「 %s 」を追加しました。',
-	'river:reply:object:groupforumtopic' => '%s さんは、議題「 %s 」に返答しました',
-	'river:reply:view' => '返答を見る',
 
 	'groups:nowidgets' => 'このグループに設定されているウィジェットはありません。',
 
@@ -280,8 +188,6 @@ return array(
 	'groups:widgets:entities:label:displaynum' => 'グループのオブジェクト一覧',
 	'groups:widgets:entities:label:pleaseedit' => 'このウィジェットを設定してください。',
 
-	'groups:forumtopic:edited' => '議題を編集しました。',
-
 	'groups:allowhiddengroups' => 'プライベート（不可視）なグループを許可しますか？',
 	'groups:whocancreate' => 'グループを新規作成できる人',
 
@@ -292,13 +198,6 @@ return array(
 	'group:notdeleted' => 'グループが削除できませんでした。',
 
 	'group:notfound' => 'そのグループは見つけることができませんでした。',
-	'grouppost:deleted' => '投稿を削除しました。',
-	'grouppost:notdeleted' => '投稿が削除できませんでした。',
-	'groupstopic:deleted' => '議題を削除しました。',
-	'groupstopic:notdeleted' => '議題を削除できませんでした。',
-	'grouptopic:blank' => '議題はありません。',
-	'grouptopic:notfound' => 'その議題は見つかりませんでした。',
-	'grouppost:nopost' => '投稿がありません。',
 	'groups:deletewarning' => "このグループを削除していいですか？削除したら元に戻す事はできません！",
 
 	'groups:invitekilled' => '招待状を削除しました。',
@@ -309,6 +208,5 @@ return array(
 	/**
 	 * ecml
 	 */
-	'groups:ecml:discussion' => 'グループ会議',
 	'groups:ecml:groupprofile' => 'グループプロフィール',
 );