+++ mod/members/languages/it.php
@@ -4,8 +4,9 @@ return array(
 	'members:search' => 'Cerca membri',
 	'members:title:search' => 'Ricerca membri per %s',
 	'members:total' => 'Totale membri: %s',
-	'members:title:newest' => 'Membri nuovissimi',
+	'members:title:newest' => 'Membri recenti',
 	'members:title:popular' => 'Membri più popolari',
 	'members:title:online' => 'Membri online',
+	'members:title:alpha' => 'Tutti i membri',
 	'members:list:popular:none' => 'Nessun membro ha nessun amico',
 );