+++ views/default/elements/forms.css.php
@@ -258,6 +258,16 @@ input[type="radio"] {
 	background-color: #eee;
 	display: block;
 }
+.ui-helper-hidden-accessible {
+	border: 0;
+	clip: rect(0 0 0 0);
+	height: 1px;
+	margin: -1px;
+	overflow: hidden;
+	padding: 0;
+	position: absolute;
+	width: 1px;
+}
 
 /* ***************************************
 	USER PICKER